fix: move cli-testing lib into new sub package
