+++ libs/cli-testing/cli-project/src/index.ts
@@ -1,1 +1,3 @@
-export * from './lib/cli-testing-cli-project';
+export * from './lib/types';
+export * from './lib/utils';
+export * from './lib/cli';
