--- libs/cli-testing/cli-project/src/lib/cli-testing-cli-project.spec.ts
@@ -1,7 +1,0 @@
-import { cliTestingCliProject } from './cli-testing-cli-project';
-
-describe('cliTestingCliProject', () => {
-  it('should work', () => {
-    expect(cliTestingCliProject()).toEqual('cli-testing-cli-project');
-  });
-});
