--- libs/cli-testing/cli-project/src/lib/cli-testing-cli-project.ts
@@ -1,3 +1,0 @@
-export function cliTestingCliProject(): string {
-  return 'cli-testing-cli-project';
-}
