--- packages/cli/tests/cli-testing/cli-project/index.ts
@@ -1,3 +1,0 @@
-export * from './types';
-export * from './utils';
-export * from './cli';
